perf(sql): run update_value_rows and recalculate_ref_types before caching
Move `update_value_rows()` and `recalculate_ref_types()` into
`resolve_metadata()` so their transformations are now cached.

XXX: This commmit enforces compatible projection types for UNION/EXCEPT
     in all queries, whereas previously this was only checked for queries
     with parameters.

+++ sql-planner/src/frontend/sql/ir/tests/global.rs
@@ -938,7 +938,7 @@ fn front_sql_global_union_all3() {
 
     insta::assert_snapshot!(plan.as_explain().unwrap(), @r#"
     union all
-        projection ("unnamed_subquery"."a"::int -> "a")
+        projection ("unnamed_subquery"."a"::decimal -> "a")
             scan "unnamed_subquery"
                 union all
                     projection ("global_t"."a"::int -> "a")
